@@ -3,7 +3,7 @@
     {
         <div class="title">
             <h2 class="my-3">@Itineraries.Count results</h2>
-            <select class="custom-select" bind=chosenSortOrder>
+            <select class="custom-select" @bind="chosenSortOrder">
                 <option value=@SortOrder.Price>Cheapest</option>
                 <option value=@SortOrder.Duration>Quickest</option>
             </select>
@@ -22,18 +22,21 @@
                 </ul>
                 <div class="price ml-3">
                     <h3>@item.Price.ToString("$0")</h3>
-                    <button class="btn" onclick=@(() => OnAddItinerary(item))>Add</button>
+                    <button class="btn" @onclick="@(() => OnAddItinerary.InvokeAsync(item))">Add</button>
                 </div>
             </div>
         }
     }
 </div>
 
-@functions
+@code
 {
     // Parameters
-    [Parameter] IReadOnlyList<Itinerary> Itineraries { get; set; }
-    [Parameter] Action<Itinerary> OnAddItinerary { get; set; }
+    [Parameter]
+    private IReadOnlyList<Itinerary> Itineraries { get; set; }
+
+    [Parameter]
+    private EventCallback<Itinerary> OnAddItinerary { get; set; }
 
     // Private state
     private SortOrder chosenSortOrder;
@@ -42,5 +45,5 @@
         ? Itineraries.OrderBy(x => x.Price)
         : Itineraries.OrderBy(x => x.TotalDurationHours);
 
-    enum SortOrder { Price, Duration }
+    private enum SortOrder { Price, Duration }
 }
