@@ -1,4 +1,4 @@
-﻿\<div class="date">
+﻿<div class="date">
     <h4>@Flight.DepartureTime.ToString("ddd MMM d")</h4>
     @Flight.TicketClass.ToDisplayString()
 </div>
@@ -15,7 +15,8 @@
     @Flight.ToAirportCode
 </div>
 
-@functions
+@code
 {
-    [Parameter] FlightSegment Flight { get; set; }
+    [Parameter]
+    private FlightSegment Flight { get; set; }
 }
