@@ -1,6 +1,7 @@
 ﻿@using System.Net.Http
 @using Microsoft.AspNetCore.Blazor
-@using Microsoft.AspNetCore.Blazor.Components
+@using Microsoft.AspNetCore.Components
+@using FlightFinder.Shared
 @using FlightFinder.Shared.Services
 @using FlightFinder.Common.Services
 @using FlightFinder.Common.Models
